+++ src/main/java/local/happysixplus/backendcodeanalysis/vo/UserVo.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class UserVo {
-    Integer id;
+    Long id;
     String username;
     String pwdMd5;
 }
